+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v1.3.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- `PyAuditError` added for non-subprocess errors
+
 ### Changed
 - Improves `pyaud format` error handling
 - Sets `Black` loglevel to debug